tests/wpt/tests/navigation-api/navigation-methods/navigate-history-push-not-loaded.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  let start_length = navigation.entries().length;
  let start_index = navigation.currentEntry.index;
  // Purposefully do not wait until after the load event (unlike some sibling tests).
  assert_equals(document.readyState, "loading", "Document must not have loaded yet");

  let navigateEventType;
  navigation.onnavigate = e => navigateEventType = e.navigationType;

  await navigation.navigate("#1", { history: "push" }).finished;
  assert_equals(navigateEventType, "push");
  assert_equals(navigation.entries().length, start_length+1);
  assert_equals(navigation.currentEntry.index, start_index+1);
}, "navigate() with history: 'push' in a document that has not yet had its load event");
</script>
